Add second MVVM UIKit template: for Request
There are certain common patterns when creating components in the app.

You have a:

- ViewModel, ViewModelDelegate
- View, ViewData, ViewDelegate
- ViewController

That are commonly used by fetching some data from the backend and when
loaded it, display it.

Then I took the initial MVVM template and add default code for this
common use.

In this setting the `ViewModel` is a delegate of the `View`

We use an `activeView` to determine which view to present for the
according states:

- loading
- dataLoaded
- failure

This uses common views like `ErrorView` and `LoadingView` to achieve
this.